rename queries to commands throughout
