+++ src/Commands/Bag.php
@@ -1,5 +1,5 @@
 <?php
-namespace Michaels\Spider\Queries;
+namespace Michaels\Spider\Commands;
 
 /**
  * Class Bag
